RetainInstance fixed comment list
