+++ app/src/main/AndroidManifest.xml
@@ -22,6 +22,7 @@
             </intent-filter>
         </activity>
         <activity android:name=".view.CommentActivity"
+            android:configChanges="orientation|screenSize"
             android:parentActivityName=".view.MainActivity"
             tools:ignore="UnusedAttribute">
             <meta-data android:name="android.support.PARENT_ACTIVITY"
