+++ app/src/main/java/com/anythingintellect/hackernews/di/ContextComponent.java
@@ -1,5 +1,6 @@
 package com.anythingintellect.hackernews.di;
 
+import com.anythingintellect.hackernews.view.CommentActivity;
 import com.anythingintellect.hackernews.view.CommentListFragment;
 import com.anythingintellect.hackernews.view.MainActivity;
 
@@ -11,7 +12,7 @@ import dagger.Subcomponent;
 public interface ContextComponent {
 
     void inject(MainActivity mainActivity);
-    void inject(CommentListFragment commentListFragment);
+    void inject(CommentActivity commentActivity);
 
 }
